refactor: change player variables to be an object for reassignment
However, there was an issue when trying to assign the player to a new instance of YT.Player in the onYouTubeIframeAPIReady function. This was causing the error: "Cannot set property player of #<Object> which has only a getter."

By creating player as an object, player can now accept new properties, or have its existing properties reassigned.